Reformat and fix image type

+++ modular_robot_physical/revolve2/modular_robot_physical/physical_interfaces/v2/_v2_physical_interface.py
@@ -184,5 +184,5 @@ class V2PhysicalInterface(PhysicalInterface):
 
         :returns: An image captured from robohatlib.
         """
-        image = self.cam.get_capture_array()
+        image = self.cam.get_capture_array().astype(np.uint8)
         return image
